+++ hi-ml-azure/src/health_azure/utils.py
@@ -9,17 +9,17 @@ import hashlib
 import json
 import logging
 import os
-
-import pandas as pd
-import param
 import re
+import tempfile
 from argparse import ArgumentParser, OPTIONAL
 from collections import defaultdict
 from itertools import islice
 from pathlib import Path
-from typing import Any, Callable, DefaultDict, Dict, List, Optional, Tuple, Type, TypeVar, Union, Set
+from typing import Any, Callable, DefaultDict, Dict, List, Optional, Set, Tuple, Type, TypeVar, Union
 
 import conda_merge
+import pandas as pd
+import param
 import ruamel.yaml
 from azureml._restclient.constants import RunStatus
 from azureml.core import Environment, Experiment, Run, Workspace, get_run
@@ -475,24 +475,21 @@ def is_private_field_name(name: str) -> bool:
 
 def determine_run_id_type(run_or_recovery_id: str) -> str:
     """
-    Determine whether a run id is of type "run id" or "run recovery id". This distinction is made
-    by checking for telltale patterns within the string. Run recovery ideas take the form "experiment_name:run_id"
-    whereas run_ids follow the pattern of a mixture of strings and decimals, separated by underscores. If the input
+    Determine whether a run id is of type "run id" or "run recovery id". Run recovery ideas take the form
+    "experiment_name:run_id". If the input
     string takes the format of a run recovery id, only the run id part will be returned. If it is a run id already,
-    it will be returned without transformation. If neither, a ValueError is raised.
+    it will be returned without transformation.
 
     :param run_or_recovery_id: The id to determine as either a run id or a run recovery id
     :return: A string representing the run id
     """
     if run_or_recovery_id is None:
         raise ValueError("Expected run_id or run_recovery_id but got None")
-    elif len(run_or_recovery_id.split(EXPERIMENT_RUN_SEPARATOR)) > 1:
+    parts = run_or_recovery_id.split(EXPERIMENT_RUN_SEPARATOR)
+    if len(parts) > 1:
         # return only the run_id, which comes after the colon
-        return run_or_recovery_id.split(EXPERIMENT_RUN_SEPARATOR)[1]
-    elif re.search(r"\d", run_or_recovery_id) and re.search('_', run_or_recovery_id):
-        return run_or_recovery_id
-    else:
-        raise ValueError("Unknown run type. Expected run_id or run_recovery id")
+        return parts[1]
+    return run_or_recovery_id
 
 
 def _find_file(file_name: str, stop_at_pythonpath: bool = True) -> Optional[Path]:
@@ -500,7 +497,7 @@ def _find_file(file_name: str, stop_at_pythonpath: bool = True) -> Optional[Path
     Recurse up the file system, starting at the current working directory, to find a file. Optionally stop when we hit
     the PYTHONPATH root (defaults to stopping).
 
-    :param file_name: The fine name of the file to find.
+    :param file_name: The file name of the file to find.
     :param stop_at_pythonpath: (Defaults to True.) Whether to stop at the PYTHONPATH root.
     :return: The path to the file, or None if it cannot be found.
     """
@@ -510,9 +507,10 @@ def _find_file(file_name: str, stop_at_pythonpath: bool = True) -> Optional[Path
             file_name: str,
             stop_at_pythonpath: bool,
             pythonpaths: List[Path]) -> Optional[Path]:
-        for child in start_at.iterdir():
-            if child.is_file() and child.name == file_name:
-                return child
+        logging.debug(f"Searching for file {file_name} in {start_at}")
+        expected = start_at / file_name
+        if expected.is_file() and expected.name == file_name:
+            return expected
         if start_at.parent == start_at or start_at in pythonpaths:
             return None
         return return_file_or_parent(start_at.parent, file_name, stop_at_pythonpath, pythonpaths)
@@ -564,7 +562,7 @@ def get_workspace(aml_workspace: Optional[Workspace] = None, workspace_config_pa
 
 def create_run_recovery_id(run: Run) -> str:
     """
-   Creates an recovery id for a run so it's checkpoints could be recovered for training/testing
+    Creates a unique ID for a run, from which the experiment name and the run ID can be re-created
 
    :param run: an instantiated run.
    :return: recovery id for a given run in format: [experiment name]:[run id]
@@ -753,16 +751,16 @@ def create_python_environment(conda_environment_file: Path,
                               docker_base_image: str = "",
                               environment_variables: Optional[Dict[str, str]] = None) -> Environment:
     """
-    Creates a description for the Python execution environment in AzureML, based on the Conda environment
-    definition files that are specified in `source_config`. If such environment with this Conda environment already
-    exists, it is retrieved, otherwise created afresh.
+    Creates a description for the Python execution environment in AzureML, based on the arguments.
+    The environment will have a name that uniquely identifies it (it is based on hashing the contents of the
+    Conda file, the docker base image, environment variables and private wheels.
 
     :param environment_variables: The environment variables that should be set when running in AzureML.
     :param docker_base_image: The Docker base image that should be used when creating a new Docker image.
     :param pip_extra_index_url: If provided, use this PIP package index to find additional packages when building
         the Docker image.
     :param workspace: The AzureML workspace to work in, required if private_pip_wheel_path is supplied.
-    :param private_pip_wheel_path: If provided, add this wheel as a private package to the AzureML workspace.
+    :param private_pip_wheel_path: If provided, add this wheel as a private package to the AzureML environment.
     :param conda_environment_file: The file that contains the Conda environment definition.
     """
     conda_dependencies = CondaDependencies(conda_dependencies_file_path=conda_environment_file)
@@ -779,20 +777,31 @@ def create_python_environment(conda_environment_file: Path,
         **(environment_variables or {})
     }
     # See if this package as a whl exists, and if so, register it with AzureML environment.
-    if workspace is not None and private_pip_wheel_path is not None:
-        if private_pip_wheel_path.is_file():
-            whl_url = Environment.add_private_pip_wheel(workspace=workspace,
-                                                        file_path=str(private_pip_wheel_path),
-                                                        exist_ok=True)
-            conda_dependencies.add_pip_package(whl_url)
-            print(f"Added add_private_pip_wheel {private_pip_wheel_path} to AzureML environment.")
-        else:
-            raise FileNotFoundError(f"Cannot add add_private_pip_wheel: {private_pip_wheel_path}, it is not a file.")
+    if private_pip_wheel_path is not None:
+        if not private_pip_wheel_path.is_file():
+            raise FileNotFoundError(f"Cannot add private wheel: {private_pip_wheel_path} is not a file.")
+        if workspace is None:
+            raise ValueError("To use a private pip wheel, an AzureML workspace must be provided.")
+        whl_url = Environment.add_private_pip_wheel(workspace=workspace,
+                                                    file_path=str(private_pip_wheel_path),
+                                                    exist_ok=True)
+        conda_dependencies.add_pip_package(whl_url)
+        logging.info(f"Added add_private_pip_wheel {private_pip_wheel_path} to AzureML environment.")
     # Create a name for the environment that will likely uniquely identify it. AzureML does hashing on top of that,
     # and will re-use existing environments even if they don't have the same name.
     # Hashing should include everything that can reasonably change. Rely on hashlib here, because the built-in
-    # hash function gives different results for the same string in different python instances.
-    hash_string = "\n".join([yaml_contents, docker_base_image, str(environment_variables)])
+    hash_string = "\n".join([yaml_contents,
+                             docker_base_image,
+                             # Changing the index URL can lead to differences in package version resolution
+                             pip_extra_index_url,
+                             str(environment_variables),
+                             # Use the path of the private wheel as a proxy. This could lead to problems if
+                             # a new environment uses the same private wheel file name, but the wheel has different
+                             # contents. In hi-ml PR builds, the wheel file name is unique to the build, so it
+                             # should not occur there.
+                             str(private_pip_wheel_path)])
+    # Python's hash function gives different results for the same string in different python instances,
+    # hence need to use hashlib
     sha1 = hashlib.sha1(hash_string.encode("utf8"))
     overall_hash = sha1.hexdigest()[:32]
     unique_env_name = f"HealthML-{overall_hash}"
@@ -1390,3 +1399,80 @@ def download_files_from_hyperdrive_children(run: Run, remote_file_path: str, loc
             downloaded_file_paths.append(str(downloaded_file_path))
 
     return downloaded_file_paths
+
+
+def create_aml_run_object(experiment_name: str,
+                          run_name: Optional[str] = None,
+                          workspace: Optional[Workspace] = None,
+                          workspace_config_path: Optional[Path] = None,
+                          snapshot_directory: Optional[PathOrString] = None) -> Run:
+    """
+    Creates an AzureML Run object in the given workspace, or in the workspace given by the AzureML config file.
+    This Run object can be used to write metrics to AzureML, upload files, etc, when the code is not running in
+    AzureML. After finishing all operations, use `run.flush()` to write metrics to the cloud, and `run.complete()` or
+    `run.fail()`.
+
+    Example:
+    >>>run = create_aml_run_object(experiment_name="run_on_my_vm", run_name="try1")
+    >>>run.log("foo", 1.23)
+    >>>run.flush()
+    >>>run.complete()
+
+    :param experiment_name: The AzureML experiment that should hold the run that will be created.
+    :param run_name: An optional name for the run (this will be used as the display name in the AzureML UI)
+    :param workspace: If provided, use this workspace to create the run in. If not provided, use the workspace
+        specified by the `config.json` file in the folder or its parent folder(s).
+    :param workspace_config_path: If not provided with an AzureML workspace, then load one given the information in this
+        config file.
+    :param snapshot_directory: The folder that should be included as the code snapshot. By default, no snapshot
+        is created (snapshot_directory=None or snapshot_directory=""). Set this to the folder that contains all the
+        code your experiment uses. You can use a file .amlignore to skip specific files or folders, akin to .gitignore
+    :return: An AzureML Run object.
+    """
+    actual_workspace = get_workspace(aml_workspace=workspace, workspace_config_path=workspace_config_path)
+    exp = Experiment(workspace=actual_workspace, name=experiment_name)
+    if snapshot_directory is None or snapshot_directory == "":
+        snapshot_directory = tempfile.mkdtemp()
+    return exp.start_logging(name=run_name, snapshot_directory=str(snapshot_directory))  # type: ignore
+
+
+def aml_workspace_for_unittests() -> Workspace:
+    """
+    Gets the default AzureML workspace that is used for unit testing. It first tries to locate a workspace config.json
+    file in the present folder or its parents, and create a workspace from that if found. If no config.json file
+    is found, the workspace details are read from environment variables. Authentication information is also read
+    from environment variables.
+    """
+    config_json = _find_file(WORKSPACE_CONFIG_JSON)
+    if config_json is not None:
+        return Workspace.from_config(path=str(config_json))
+    else:
+        workspace_name = get_secret_from_environment(ENV_WORKSPACE_NAME, allow_missing=False)
+        subscription_id = get_secret_from_environment(ENV_SUBSCRIPTION_ID, allow_missing=False)
+        resource_group = get_secret_from_environment(ENV_RESOURCE_GROUP, allow_missing=False)
+        auth = get_authentication()
+        return Workspace.get(name=workspace_name,
+                             auth=auth,
+                             subscription_id=subscription_id,
+                             resource_group=resource_group)
+
+
+class UnitTestWorkspaceWrapper:
+    """
+    Wrapper around aml_workspace so that it is lazily loaded only once. Used for unit testing only.
+    """
+
+    def __init__(self) -> None:
+        """
+        Init.
+        """
+        self._workspace: Workspace = None
+
+    @property
+    def workspace(self) -> Workspace:
+        """
+        Lazily load the aml_workspace.
+        """
+        if self._workspace is None:
+            self._workspace = aml_workspace_for_unittests()
+        return self._workspace