+++ hi-ml-azure/testazure/testazure/utils_testazure.py
@@ -13,10 +13,8 @@ from contextlib import contextmanager
 from pathlib import Path
 from typing import Dict, Generator, Optional
 
-from azureml.core import Workspace
-
-from health_azure.utils import (ENV_RESOURCE_GROUP, ENV_SUBSCRIPTION_ID, ENV_WORKSPACE_NAME, get_authentication,
-                                get_secret_from_environment, WORKSPACE_CONFIG_JSON)
+from health_azure.utils import (ENV_RESOURCE_GROUP, ENV_SUBSCRIPTION_ID, ENV_WORKSPACE_NAME, WORKSPACE_CONFIG_JSON,
+                                UnitTestWorkspaceWrapper)
 
 DEFAULT_DATASTORE = "himldatasets"
 FALLBACK_SINGLE_RUN = "refs_pull_545_merge:refs_pull_545_merge_1626538212_d2b07afd"
@@ -50,46 +48,7 @@ def repository_root() -> Path:
     return himl_azure_root().parent
 
 
-def default_aml_workspace() -> Workspace:
-    """
-    Gets the default AzureML workspace that is used for testing.
-    """
-    config_json = repository_root() / WORKSPACE_CONFIG_JSON
-    if config_json.is_file():
-        return Workspace.from_config()
-    else:
-        workspace_name = get_secret_from_environment(ENV_WORKSPACE_NAME, allow_missing=False)
-        subscription_id = get_secret_from_environment(ENV_SUBSCRIPTION_ID, allow_missing=False)
-        resource_group = get_secret_from_environment(ENV_RESOURCE_GROUP, allow_missing=False)
-        auth = get_authentication()
-        return Workspace.get(name=workspace_name,
-                             auth=auth,
-                             subscription_id=subscription_id,
-                             resource_group=resource_group)
-
-
-class WorkspaceWrapper:
-    """
-    Wrapper around aml_workspace so that it is lazily loaded, once.
-    """
-
-    def __init__(self) -> None:
-        """
-        Init.
-        """
-        self._workspace: Workspace = None
-
-    @property
-    def workspace(self) -> Workspace:
-        """
-        Lazily load the aml_workspace.
-        """
-        if self._workspace is None:
-            self._workspace = default_aml_workspace()
-        return self._workspace
-
-
-DEFAULT_WORKSPACE = WorkspaceWrapper()
+DEFAULT_WORKSPACE = UnitTestWorkspaceWrapper()
 
 
 @contextmanager
@@ -104,6 +63,13 @@ def change_working_directory(path_or_str: Path) -> Generator:
     os.chdir(old_path)
 
 
+def get_shared_config_json() -> Path:
+    """
+    Gets the path to the config.json file that should exist for running tests locally (outside github build agents).
+    """
+    return repository_root() / "hi-ml-azure" / "testazure" / WORKSPACE_CONFIG_JSON
+
+
 @contextmanager
 def check_config_json(script_folder: Path) -> Generator:
     """
@@ -111,20 +77,27 @@ def check_config_json(script_folder: Path) -> Generator:
     from the repository root folder (this should be the case when executing a test on a dev machine), or create
     it from environment variables (this should trigger in builds on the github agents).
     """
-    shared_config_json = repository_root() / WORKSPACE_CONFIG_JSON
+    shared_config_json = get_shared_config_json()
     target_config_json = script_folder / WORKSPACE_CONFIG_JSON
     if shared_config_json.exists():
         logging.info(f"Copying {WORKSPACE_CONFIG_JSON} from repository root to folder {script_folder}")
         shutil.copy(shared_config_json, target_config_json)
     else:
         logging.info(f"Creating {str(target_config_json)} from environment variables.")
-        with open(str(target_config_json), 'w', encoding="utf-8") as file:
-            config = {
-                "subscription_id": os.getenv(ENV_SUBSCRIPTION_ID, ""),
-                "resource_group": os.getenv(ENV_RESOURCE_GROUP, ""),
-                "workspace_name": os.getenv(ENV_WORKSPACE_NAME, "")
-            }
-            json.dump(config, file)
+        subscription_id = os.getenv(ENV_SUBSCRIPTION_ID, "")
+        resource_group = os.getenv(ENV_RESOURCE_GROUP, "")
+        workspace_name = os.getenv(ENV_WORKSPACE_NAME, "")
+        if subscription_id and resource_group and workspace_name:
+            with open(str(target_config_json), 'w', encoding="utf-8") as file:
+                config = {
+                    "subscription_id": os.getenv(ENV_SUBSCRIPTION_ID, ""),
+                    "resource_group": os.getenv(ENV_RESOURCE_GROUP, ""),
+                    "workspace_name": os.getenv(ENV_WORKSPACE_NAME, "")
+                }
+                json.dump(config, file)
+        else:
+            raise ValueError("Either a shared config.json must be present, or all 3 environment variables for "
+                             "workspace creation must exist.")
     try:
         yield
     finally: